Adding management of vacancies, resources and reviews to the administrator panel
