+++ linkit-project-client/src/App.tsx
@@ -25,8 +25,8 @@ import {
 import { motion, Variants } from "framer-motion";
 import axios, { AxiosError } from "axios";
 import { useEffect } from "react";
-import AdminPanel from "./components/Paneles/AdminProfile/Panel/AdminPanel.tsx";
 import { SUPERADMN_ID } from "./env.ts";
+import AdminPanel from "./components/Paneles/AdminProfile/Panel/AdminPanel.tsx";
 
 const registerVariants: Variants = {
   hidden: {
@@ -173,7 +173,7 @@ function App() {
         <Route path="/recursos" element={<Recursos />} />
         <Route path="/recursos/libreria" element={<Libreria />} />
         <Route path="/quienesSomos" element={<QuienesSomos />} />
-        <Route path="/AdminDashboard" element={<AdminPanel />} />
+        <Route path="/AdminDashboard/*" element={<AdminPanel />} />
       </Routes>
     </>
   );
