+++ linkit-project-client/src/components/Paneles/AdminProfile/Panel/AdminRecursos/AdminRecursos.tsx
@@ -1,0 +1,7 @@
+export default function AdminRecursos() {
+  return (
+    <div>
+        <h1 className="text-5xl pl-32 pb-4">Gestión de recursos</h1>
+    </div>
+  )
+}
