+++ linkit-project-client/src/components/Paneles/AdminProfile/Panel/AdminReviews/AdminReviews.tsx
@@ -1,0 +1,9 @@
+
+
+export default function AdminReviews() {
+  return (
+    <div>
+        <h1 className="text-5xl pl-32 pb-4">Gestión de reviews</h1>
+    </div>
+  )
+}
